HighSpeedActivity: support recording in HDR
Support recording high frame rates in HDR by requesting the recording
surface to use BT.2020 and HLG. Since the MediaRecorder API seems to
not allow setting these values, the HDR case uses a combination of
the low-level MediaCodec interface together with MediaRecorder.
Recording in HDR uses HEVC, because H.264 (or most implementations)
do not support 10-bit encoded content well.

This functionality is not particularly well-tested. It is up to the
Camera HAL to respect the requested parameters (or to override them),
making HDR support a very device-specific feature.